added README.md refactor code
