+++ test/conformance/kuttl/cleanup/ttl-cleanup/valid-label/README.md
@@ -1,0 +1,9 @@
+# ## Description
+
+This test cleans up pods via a label assignment named `kyverno.io/ttl: 10s`.
+
+## Expected Behavior
+
+The pod `test-pod` is cleaned up successfully after 10s.
+
+## Reference Issue(s)
